+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- PSR-14 event for modifying site configuration on runtime (#29)
+
 ## [2.0.0] - 2024-02-09
 
 ### Added